Extract UIImage test helper to a separate file and project organization
