+++ MyEssentialFeed/MyEssentialFeediOS/Feed Image Loader/FeedImageDataLoader.swift
@@ -1,0 +1,17 @@
+//
+//  FeedImageDataLoader.swift
+//  MyEssentialFeediOS
+//
+//  Created by Tiago Chaves on 15/01/24.
+//
+
+import Foundation
+
+public protocol FeedImageDataLoaderTask {
+    func cancel()
+}
+
+public protocol FeedImageDataLoader {
+    typealias Result = Swift.Result<Data, Error>
+    func loadImageData(from url: URL, completion: @escaping (Result) -> Void) -> FeedImageDataLoaderTask
+}
